Reorganize SideNav into its own folder
