+++ app/javascript/components/common/SideNav/Logo.tsx
