+++ app/javascript/packs/side-nav.tsx
@@ -1,6 +1,6 @@
 // License: LGPL-3.0-or-later
 import ReactOnRails from 'react-on-rails';
-import SideNav from '../components/common/SideNavRenderFunc';
+import SideNav from '../components/common/SideNav/SideNavRenderFunc';
 
 // This is how react_on_rails can see SignInPage in the browser.
 ReactOnRails.register({
